fix(permissions): strip literal '*' from always-allow prefix in Desktop/ACP
Desktop and ACP were sending prefixes like \"dotnet *\" to the backend,
but isAllowed() does exact token matching, so the literal '*' never
matched any command. This made \"always allow\" silently fail.

- Move derivePrefix() to src/tools/shell/parse.ts so CLI/ACP can share it
- Desktop inlines the same derivePrefix() logic (separate build target)
- Fixes #1180
